ADD: construction of the sith workflow
This workflow starts creating the peptide and equilibrating it
then, pull it via classical MD. From that trajectory, the most
stretched configuration is taken and optimized with gaussian to
obtain the local minimum corresponding to the stretched config.
Finally, that configuration is restretched by constrainig an in-
creased distance between C-cap atoms and optimizing with that
constrain.

This workflow still needs to consider the sith analysis.