DEV: package behaviour improved
There is still a problem with the installation because
it is not importing the directories ase_utils and gromacs
for some reason. Surely it will be next commit.

Also, some functions don't have the flag add2executable
and then cannot be executed from the terminal. This has
to be done before commiting the next myutils version