ENH: unification of command line for the function that changes the distance in the stretching
With this change, all the functions created to increase the
distance between two atoms of the same molecule have frozen
dofs as an argument instead of indexes. In this way all the
functions have the same arguments and only one command line
is necessary to execute any method, defining the later as
another argument.